+++ ldaporm/types.py
